refactor(test): 리플렉션으로 작성한 엔티티 Mockito로 수정
