+++ src/test/java/com/skhu/gdgocteambuildingproject/projectgallery/model/GalleryProjectAccessCheckerTest.java
@@ -10,13 +10,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.springframework.test.util.ReflectionTestUtils;
+import org.springframework.security.core.Authentication;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-
-import java.lang.reflect.Constructor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
@@ -40,17 +38,14 @@ class GalleryProjectAccessCheckerTest {
     @Test
     void User_엔티티의_Admin_권한을_체크한다() throws Exception {
         // given
-        Constructor<User> constructor = User.class.getDeclaredConstructor();
-        constructor.setAccessible(true);
-        User admin = constructor.newInstance();
+        User mockUser = mock(User.class);
 
-        ReflectionTestUtils.setField(admin, "role", ADMIN_ROLE);
+        when(mockUser.getRole()).thenReturn(ADMIN_ROLE);
+
+        Authentication authentication = new UsernamePasswordAuthenticationToken(mockUser, null, null);
 
         // when
-        boolean result = checker.checkLeaderOrAdminPermission(
-                PROJECT_ID,
-                new UsernamePasswordAuthenticationToken(admin, null, null)
-        );
+        boolean result = checker.checkLeaderOrAdminPermission(PROJECT_ID, authentication);
 
         // then
         assertThat(result).isTrue();
@@ -60,21 +55,18 @@ class GalleryProjectAccessCheckerTest {
     @Test
     void User_엔티티가_프로젝트_리더인지_체크한다() throws Exception {
         // given
-        Constructor<User> constructor = User.class.getDeclaredConstructor();
-        constructor.setAccessible(true);
-        User leader = constructor.newInstance();
+        User mockUser = mock(User.class);
 
-        ReflectionTestUtils.setField(leader, "id", USER_ID);
-        ReflectionTestUtils.setField(leader, "role", MEMBER_ROLE);
+        when(mockUser.getId()).thenReturn(USER_ID);
+        when(mockUser.getRole()).thenReturn(MEMBER_ROLE);
+
+        Authentication authentication = new UsernamePasswordAuthenticationToken(mockUser, null, null);
 
         when(repository.existsByProjectIdAndUserIdAndRole(PROJECT_ID, USER_ID, LEADER_ROLE))
                 .thenReturn(true);
 
         // when
-        boolean result = checker.checkLeaderOrAdminPermission(
-                PROJECT_ID,
-                new UsernamePasswordAuthenticationToken(leader, null, null)
-        );
+        boolean result = checker.checkLeaderOrAdminPermission(PROJECT_ID, authentication);
 
         // then
         assertThat(result).isTrue();
@@ -83,21 +75,18 @@ class GalleryProjectAccessCheckerTest {
     @Test
     void User_엔티티가_프로젝트의_리더가_아니면_거부한다() throws Exception {
         // given
-        Constructor<User> constructor = User.class.getDeclaredConstructor();
-        constructor.setAccessible(true);
-        User member = constructor.newInstance();
+        User mockUser = mock(User.class);
 
-        ReflectionTestUtils.setField(member, "id", USER_ID);
-        ReflectionTestUtils.setField(member, "role", MEMBER_ROLE);
+        when(mockUser.getId()).thenReturn(USER_ID);
+        when(mockUser.getRole()).thenReturn(MEMBER_ROLE);
+
+        Authentication authentication = new UsernamePasswordAuthenticationToken(mockUser, null, null);
 
         when(repository.existsByProjectIdAndUserIdAndRole(PROJECT_ID, USER_ID, LEADER_ROLE))
                 .thenReturn(false);
 
         // when
-        boolean result = checker.checkLeaderOrAdminPermission(
-                PROJECT_ID,
-                new UsernamePasswordAuthenticationToken(member, null, null)
-        );
+        boolean result = checker.checkLeaderOrAdminPermission(PROJECT_ID, authentication);
 
         // then
         assertThat(result).isFalse();
@@ -107,17 +96,14 @@ class GalleryProjectAccessCheckerTest {
     @EnumSource(value = UserRole.class, names = {"OTHERS", "BANNED"})
     void User_엔티티의_권한이_OTHERS_또는_BANNED_인지_검증한다(UserRole role) throws Exception {
         // given
-        Constructor<User> constructor = User.class.getDeclaredConstructor();
-        constructor.setAccessible(true);
-        User user = constructor.newInstance();
+        User mockUser = mock(User.class);
 
-        ReflectionTestUtils.setField(user, "role", role);
+        when(mockUser.getRole()).thenReturn(role);
+
+        Authentication authentication = new UsernamePasswordAuthenticationToken(mockUser, null, null);
 
         // when
-        boolean result = checker.checkLeaderOrAdminPermission(
-                PROJECT_ID,
-                new UsernamePasswordAuthenticationToken(user, null, null)
-        );
+        boolean result = checker.checkLeaderOrAdminPermission(PROJECT_ID, authentication);
 
         // then
         assertThat(result).isFalse();
